requirements.txt
boto3==1.34.69
botocore==1.34.69
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Deprecated==1.2.14
Faker==24.3.0
fonttools==4.50.0
idna==3.6
jmespath==1.0.1
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
opensearch-py==2.5.0
opensearch-py-ml==1.1.0
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.1
s3transfer==0.10.1
six==1.16.0
tzdata==2024.1
urllib3==1.26.18
wrapt==1.16.0
pytest==8.2.2
tenacity==8.5.0
requests_aws4auth==1.2.3
overrides==7.4.0
python-dotenv==1.0.1
